chore: copy across nse docstring and kge implementation in hydro_impl
